리뷰 저장 오류 수정

+++ schemas.py
@@ -175,7 +175,7 @@ class MovieListResponse(BaseSchema):
 # ============================================
 
 class ReviewBase(BaseSchema):
-    rating: Decimal = Field(..., ge=0.5, le=5.0, description="0.5~5.0, 0.5 단위")
+    rating: float = Field(..., ge=0.5, le=5.0, description="0.5~5.0, 0.5 단위")
     content: Optional[str] = None
     keywords: List[str] = []
     is_public: bool = True
@@ -186,7 +186,7 @@ class ReviewCreate(ReviewBase):
 
 
 class ReviewUpdate(BaseSchema):
-    rating: Optional[Decimal] = Field(None, ge=0.5, le=5.0)
+    rating: Optional[float] = Field(None, ge=0.5, le=5.0)
     content: Optional[str] = None
     keywords: Optional[List[str]] = None
     is_public: Optional[bool] = None
